requirements.txt
python-dotenv
python-binance
coinbase-advanced-py
krakenex
bitfinex-api-py